Add .enablePoll() and .disablePoll().
Indiscriminately polling everything we find didn't work so well, I guess
some values just don't like it.  Leave it to the user to poll values
that they care about, giving an example of binary and multi-level
switches to start with.

We need to revisit this at some point to correctly log whether a value
is being polled or not.

Update docs.